+++ apps/mobile/src/features/player/components/PlayerControls.tsx
@@ -12,6 +12,7 @@ import { AppState, StyleSheet, View } from 'react-native'
 import { IconButton, Tooltip, useTheme } from 'react-native-paper'
 
 import useCurrentTrack from '@/hooks/player/useCurrentTrack'
+import { analyticsService } from '@/lib/services/analyticsService'
 import * as Haptics from '@/utils/haptics'
 import logInstance from '@/utils/log'
 
@@ -96,6 +97,7 @@ export function MainPlaybackControls({
 					onInteraction?.()
 					void Haptics.performHaptics(Haptics.AndroidHaptics.Context_Click)
 					void Orpheus.skipToPrevious()
+					void analyticsService.logPlayerAction('skip_prev')
 				}}
 				testID='player-prev'
 			/>
@@ -109,11 +111,10 @@ export function MainPlaybackControls({
 					logInstance.debug('isPlaying', isPlaying)
 					if (isPlaying) {
 						await Orpheus.pause()
+						void analyticsService.logPlayerAction('pause')
 					} else {
-						// 或许可以解决 play 无响应的问题？
-						// 好吧并不能，我是小丑
-						await Orpheus.pause()
 						await Orpheus.play()
+						void analyticsService.logPlayerAction('play')
 					}
 				}}
 				mode='contained'
@@ -126,6 +127,7 @@ export function MainPlaybackControls({
 					onInteraction?.()
 					void Haptics.performHaptics(Haptics.AndroidHaptics.Context_Click)
 					void Orpheus.skipToNext()
+					void analyticsService.logPlayerAction('skip_next')
 				}}
 				testID='player-next'
 			/>
@@ -174,6 +176,9 @@ export function PlayerControls({ onOpenQueue }: { onOpenQueue: () => void }) {
 								? Orpheus.setShuffleMode(false)
 								: Orpheus.setShuffleMode(true))
 							await refetchShuffleMode()
+							void analyticsService.logPlayerAction('shuffle', {
+								mode: !shuffleMode,
+							})
 						}}
 						testID='player-mode-shuffle'
 					/>
@@ -203,6 +208,9 @@ export function PlayerControls({ onOpenQueue }: { onOpenQueue: () => void }) {
 										: RepeatMode.OFF
 							void Orpheus.setRepeatMode(nextMode)
 							setRepeatMode(nextMode)
+							void analyticsService.logPlayerAction('repeat', {
+								mode: nextMode,
+							})
 						}}
 						testID='player-mode-repeat'
 					/>
@@ -231,6 +239,7 @@ export function PlayerControls({ onOpenQueue }: { onOpenQueue: () => void }) {
 						onPress={() => {
 							void Haptics.performHaptics(Haptics.AndroidHaptics.Context_Click)
 							onOpenQueue()
+							void analyticsService.logPlayerQueueAction('open_queue')
 						}}
 						testID='player-open-queue'
 					/>